major update to many of the template files

+++ src/IHEWAcollect/templates/USGS/SSEBop.py
@@ -7,13 +7,13 @@ import datetime
 # General modules
 import os
 import sys
-
+import inspect
 import numpy as np
 import pandas as pd
 import requests
 
 # from requests.auth import HTTPBasicAuth
-# from joblib import Parallel, delayed
+from joblib import Parallel, delayed
 
 
 # IHEWAcollect Modules
@@ -40,6 +40,12 @@ def _init(status, conf):
     # From download.py
     __this.status = status
     __this.conf = conf
+    __this.path = os.path.join(
+        os.getcwd(),
+        os.path.dirname(
+            inspect.getfile(
+                inspect.currentframe()))
+    )
 
     account = conf['account']
     folder = conf['folder']
@@ -185,35 +191,54 @@ def download_product(latlim, lonlim, dates,
     status_cod = -1
     # total = len(dates)
 
-    # Create Waitbar
-    # amount = 0
-    # if is_waitbar == 1:
-    #     amount = 0
-    #     collect.WaitBar(amount, total,
-    #                     prefix='Progress:', suffix='Complete',
-    #                     length=50)
+    cores = -1
 
-    for date in dates:
-        args = get_download_args(latlim, lonlim, date,
-                                 account, folder, product)
+    arg1 = product_details(latlim, lonlim, account, product)
+    if not cores:
+        for date in dates:
+            arg2 = get_download_args(latlim, lonlim, date,
+                                     folder, product)
+           
+            args = arg2+arg1
+            
+            status_cod = start_download(args)
 
-        status_cod = start_download(args)
-
-        # Update waitbar
-        # if is_waitbar == 1:
-        #     amount += 1
-        #     collect.WaitBar(amount, total,
-        #                     prefix='Progress:', suffix='Complete',
-        #                     length=50)
-
+            # Update waitbar
+            # if is_waitbar == 1:
+            #     amount += 1
+            #     collect.WaitBar(amount, total,
+            #                     prefix='Progress:', suffix='Complete',
+            #                     length=50)
+    else:
+        status_cod = Parallel(n_jobs=cores)(
+            delayed(start_download)(get_download_args(
+                    latlim, lonlim, date,
+                    folder, product)+arg1) for date in dates)
+        
     return status_cod
+    # # Create Waitbar
+    # # amount = 0
+    # # if is_waitbar == 1:
+    # #     amount = 0
+    # #     collect.WaitBar(amount, total,
+    # #                     prefix='Progress:', suffix='Complete',
+    # #                     length=50)
 
+    # for date in dates:
+    #     args = get_download_args(latlim, lonlim, date,
+    #                              account, folder, product)
 
-def get_download_args(latlim, lonlim, date,
-                      account, folder, product) -> tuple:
-    msg = 'Collecting  "{f}"'.format(f=date)
-    print('\33[95m{}\33[0m'.format(msg))
-    __this.Log.write(datetime.datetime.now(), msg=msg)
+    #     status_cod = start_download(args)
+
+    #     # Update waitbar
+    #     # if is_waitbar == 1:
+    #     #     amount += 1
+    #     #     collect.WaitBar(amount, total,
+    #     #                     prefix='Progress:', suffix='Complete',
+    #     #                     length=50)
+
+    # return status_cod
+def product_details(latlim, lonlim, account, product)-> tuple:
 
     # Define arg_account
     try:
@@ -228,7 +253,50 @@ def get_download_args(latlim, lonlim, date,
     # Define arg_url
     url_server = product['url']
 
-    # url_dir
+    data_ndv = product['nodata']
+    data_type = product['data']['dtype']['l']
+    data_multiplier = float(product['data']['units']['m'])
+    data_variable = product['data']['variable']
+
+    # Define arg_IDs
+    prod_lon_w = product['data']['lon']['w']
+    prod_lat_n = product['data']['lat']['n']
+    # prod_lon_e = product['data']['lon']['e']
+    # prod_lat_s = product['data']['lat']['s']
+    prod_lat_size = abs(product['data']['lat']['r'])
+    prod_lon_size = abs(product['data']['lon']['r'])
+
+    # Define arg_GTiff
+    pixel_h = int(product['data']['dem']['h'])
+    pixel_w = int(product['data']['dem']['w'])
+    pixel_size = max(prod_lat_size, prod_lon_size)
+
+    # Calculate arg_IDs
+  
+    y_id = np.array([
+        np.floor((prod_lat_n - latlim[1]) / prod_lat_size),
+        np.ceil((prod_lat_n - latlim[0]) / prod_lat_size)
+    ], dtype=np.int)
+    x_id = np.array([
+        np.floor((lonlim[0] - prod_lon_w) / prod_lon_size),
+        np.ceil((lonlim[1] - prod_lon_w) / prod_lon_size)
+    ], dtype=np.int)
+
+
+    return  product, \
+        username, password, apitoken, \
+        url_server,y_id, x_id, pixel_size, pixel_w, pixel_h, \
+        data_ndv, data_type, data_multiplier, data_variable, __this.path, __this.conf['is_save_list'], __this.Log,\
+        __this.conf
+
+
+def get_download_args(latlim, lonlim, date,
+                     folder, product) -> tuple:
+    msg = 'Collecting  "{f}"'.format(f=date)
+    print('\33[95m{}\33[0m'.format(msg))
+    __this.Log.write(datetime.datetime.now(), msg=msg)
+
+
     fmt_d = product['data']['fmt']['d']
     if fmt_d is None:
         if product['data']['dir'] is None:
@@ -294,10 +362,7 @@ def get_download_args(latlim, lonlim, date,
     pixel_w = int(product['data']['dem']['w'])
     pixel_size = max(prod_lat_size, prod_lon_size)
 
-    # Calculate arg_IDs
-    # [w,n]--[e,n]
-    #   |      |
-    # [w,s]--[e,s]
+
     y_id = np.array([
         np.floor((prod_lat_n - latlim[1]) / prod_lat_size),
         np.ceil((prod_lat_n - latlim[0]) / prod_lat_size)
@@ -307,64 +372,44 @@ def get_download_args(latlim, lonlim, date,
         np.ceil((lonlim[1] - prod_lon_w) / prod_lon_size)
     ], dtype=np.int)
 
-    # [w,s]--[e,s]
-    #   |      |
-    # [w,n]--[e,n]
-    # y_id = np.array([
-    #     np.floor((latlim[0] - prod_lat_s) / prod_lat_size),
-    #     np.ceil((latlim[1] - prod_lat_s) / prod_lat_size)
-    # ], dtype=np.int)
-    # x_id = np.array([
-    #     np.floor((lonlim[0] - prod_lon_w) / prod_lon_size),
-    #     np.ceil((lonlim[1] - prod_lon_w) / prod_lon_size)
-    # ], dtype=np.int)
-
-    # [w,n]--[w,s]
-    #   |      |
-    # [e,n]--[e,s]
-    # y_id = np.array([
-    #     np.floor((lonlim[0] - prod_lon_w) / prod_lon_size),
-    #     np.ceil((lonlim[1] - prod_lon_w) / prod_lon_size)
-    # ], dtype=np.int)
-    # x_id = np.array([
-    #     np.floor((prod_lat_n - latlim[1]) / prod_lat_size),
-    #     np.ceil((prod_lat_n - latlim[0]) / prod_lat_size)
-    # ], dtype=np.int)
-
-    # [w,s]--[w,n]
-    #   |      |
-    # [e,s]--[e,n]
-    # y_id = np.array([
-    #     np.floor((lonlim[0] - prod_lon_w) / prod_lon_size),
-    #     np.ceil((lonlim[1] - prod_lon_w) / prod_lon_size)
-    # ], dtype=np.int)
-    # x_id = np.array([
-    #     np.floor((latlim[0] - prod_lat_s) / prod_lat_size),
-    #     np.ceil((latlim[1] - prod_lat_s) / prod_lat_size)
-    # ], dtype=np.int)
+    # Adjust the lon, lat limits based on the grids of the data
+    lonlim = np.array([
+        (x_id[0] * prod_lat_size + prod_lon_w),
+        (x_id[1] * prod_lat_size + prod_lon_w)
+    ], dtype=np.float)
+    latlim = np.array([
+        (prod_lat_n - y_id[1] * prod_lat_size),
+        (prod_lat_n - y_id[0] * prod_lat_size)
+    ], dtype=np.float)
 
     return latlim, lonlim, date, \
-        product, \
-        username, password, apitoken, \
-        url_server, url_dir, \
+        url_dir, \
         fname_r, fname_t, fname_l, \
-        file_r, file_t, file_l,\
-        y_id, x_id, pixel_size, pixel_w, pixel_h, \
-        data_ndv, data_type, data_multiplier, data_variable
+        file_r, file_t, file_l
 
 
 def start_download(args) -> int:
     """Retrieves data
     """
     # Unpack the arguments
+
     latlim, lonlim, date, \
+        url_dir, \
+        remote_fname, temp_fname, local_fname, \
+        remote_file, temp_file, local_file, \
         product, \
         username, password, apitoken, \
-        url_server, url_dir, \
-        remote_fname, temp_fname, local_fname,\
-        remote_file, temp_file, local_file,\
-        y_id, x_id, pixel_size, pixel_w, pixel_h, \
-        data_ndv, data_type, data_multiplier, data_variable = args
+        url_server,y_id, x_id, pixel_size, pixel_w, pixel_h, \
+        data_ndv, data_type, data_multiplier, data_variable, current_path, save_list, log_file, current_conf = args
+
+    # latlim, lonlim, date, \
+    #     product, \
+    #     username, password, apitoken, \
+    #     url_server, url_dir, \
+    #     remote_fname, temp_fname, local_fname,\
+    #     remote_file, temp_file, local_file,\
+    #     y_id, x_id, pixel_size, pixel_w, pixel_h, \
+    #     data_ndv, data_type, data_multiplier, data_variable, current_path, save_list, log_file,current_conf = args
 
     # Define local variable
     status_cod = -1
@@ -378,13 +423,13 @@ def start_download(args) -> int:
 
             msg = 'Exist "{f}"'.format(f=local_file)
             print('\33[92m{}\33[0m'.format(msg))
-            __this.Log.write(datetime.datetime.now(), msg=msg)
+            log_file.write(datetime.datetime.now(), msg=msg)
 
     if is_start_download:
         # Download the data from server if the file not exists
         msg = 'Downloading "{f}"'.format(f=remote_fname)
         print('{}'.format(msg))
-        __this.Log.write(datetime.datetime.now(), msg=msg)
+        log_file.write(datetime.datetime.now(), msg=msg)
 
         is_download = True
         if os.path.exists(remote_file):
@@ -393,7 +438,7 @@ def start_download(args) -> int:
 
                 msg = 'Exist "{f}"'.format(f=remote_file)
                 print('\33[93m{}\33[0m'.format(msg))
-                __this.Log.write(datetime.datetime.now(), msg=msg)
+                log_file.write(datetime.datetime.now(), msg=msg)
 
         # ------------- #
         # Download data #
@@ -422,7 +467,7 @@ def start_download(args) -> int:
                     dr=url_dir,
                     fn=remote_fname)
                 print('\33[91m{}\n{}\33[0m'.format(msg, str(err)))
-                __this.Log.write(datetime.datetime.now(),
+                log_file.write(datetime.datetime.now(),
                                  msg='{}\n{}'.format(msg, str(err)))
                 remote_file_status += 1
             else:
@@ -453,7 +498,7 @@ def start_download(args) -> int:
     status_cod = remote_file_status + local_file_status
 
     msg = 'Finish'
-    __this.Log.write(datetime.datetime.now(), msg=msg)
+    log_file.write(datetime.datetime.now(), msg=msg)
     return status_cod
 
 
@@ -461,14 +506,22 @@ def convert_data(args):
     """
     """
     # Unpack the arguments
-    latlim, lonlim, date,\
+    latlim, lonlim, date, \
+        url_dir, \
+        remote_fname, temp_fname, local_fname, \
+        remote_file, temp_file, local_file, \
         product, \
         username, password, apitoken, \
-        url_server, url_dir, \
-        remote_fname, temp_fname, local_fname,\
-        remote_file, temp_file, local_file,\
-        y_id, x_id, pixel_size, pixel_w, pixel_h, \
-        data_ndv, data_type, data_multiplier, data_variable = args
+        url_server,y_id, x_id, pixel_size, pixel_w, pixel_h, \
+        data_ndv, data_type, data_multiplier, data_variable, current_path, save_list, log_file,current_conf = args
+    # latlim, lonlim, date,\
+    #     product, \
+    #     username, password, apitoken, \
+    #     url_server, url_dir, \
+    #     remote_fname, temp_fname, local_fname,\
+    #     remote_file, temp_file, local_file,\
+    #     y_id, x_id, pixel_size, pixel_w, pixel_h, \
+    #     data_ndv, data_type, data_multiplier, data_variable, current_path, save_list, log_file,current_conf = args
 
     # Define local variable
     status_cod = -1
@@ -478,7 +531,7 @@ def convert_data(args):
     #   -> local (gis)
     msg = 'Converting  "{f}"'.format(f=local_file)
     print('\33[94m{}\33[0m'.format(msg))
-    __this.Log.write(datetime.datetime.now(), msg=msg)
+    log_file.write(datetime.datetime.now(), msg=msg)
 
     # --------- #
     # Load data #
@@ -553,31 +606,34 @@ def convert_data(args):
            latlim[1], 0, -pixel_size]
     Save_as_tiff(name=local_file, data=data, geo=geo, projection="WGS84")
 
-    if __this.conf['is_save_remote']:
+    part_file_name = 'm{dtime:%Y%m}'.format(dtime=date)
+    if current_conf['is_save_remote']:
         pass
     else:
         path = os.path.dirname(os.path.realpath(remote_file))
         if 'remote' != path[-6:]:
             path = os.path.join(path, 'remote')
-        clean(path)
-    if __this.conf['is_save_temp']:
+        clean(path,log_file,part_file_name)
+
+    if current_conf['is_save_temp']:
         pass
     else:
         path = os.path.dirname(os.path.realpath(temp_file))
         if 'temporary' != path[-9:]:
             path = os.path.join(path, 'temporary')
-        clean(path)
+        clean(path,log_file,part_file_name)
 
     status_cod = 0
     return status_cod
 
-
-def clean(path):
+def clean(path,log_file,part_file_name):
     msg = 'Cleaning    "{f}"'.format(f=path)
-    print('{}'.format(msg))
-    __this.Log.write(datetime.datetime.now(), msg=msg)
+    # print('{}'.format(msg))
+    log_file.write(datetime.datetime.now(), msg=msg)
 
     for root, dirs, files in os.walk(path):
         for filename in files:
             # print(filename)
-            os.remove(os.path.join(root, filename))
+            if(part_file_name in filename):
+                os.remove(os.path.join(root, filename))
+
